requirements.txt
#torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
from_root==1.0.2
tqdm==4.64.1
split-folders==0.5.1
boto3==1.24.73
pandas
scikit-learn
pymongo==4.2.0
annoy==1.17.1
dnspython
pillow
python-dotenv
-e .